+++ src/lib/vm/vm.js
@@ -2,6 +2,7 @@ import React from "react";
 import { Widget } from "../components/Widget";
 import {
   deepCopy,
+  deepEqual,
   deepFreeze,
   ipfsUpload,
   ipfsUrl,
@@ -60,6 +61,7 @@ import * as Toolbar from "@radix-ui/react-toolbar";
 import * as RadixTooltip from "@radix-ui/react-tooltip";
 import { ethers } from "ethers";
 import { Web3ConnectButton } from "../components/ethers";
+import { isFunction } from "react-bootstrap-typeahead/types/utils";
 
 const frozenNacl = Object.freeze({
   randomBytes: deepFreeze(nacl.randomBytes),
@@ -434,6 +436,7 @@ class Stack {
 class VmStack {
   constructor(vm, prevStack, state, isTrusted) {
     this.gIndex = 0;
+    this.hookIndex = 0;
     this.vm = vm;
     this.isTrusted = !!isTrusted;
     this.stack = new Stack(prevStack, state);
@@ -1029,6 +1032,75 @@ class VmStack {
             );
           }
           return this.vm.useCache(...args);
+        } else if (callee === "useState") {
+          if (this.prevStack) {
+            throw new Error(
+              "Method: useState. The hook can an only be called from the top of the stack"
+            );
+          }
+          if (args.length < 1) {
+            throw new Error(
+              "Method: useState. Required arguments: 'initialState'"
+            );
+          }
+          const initialState = args[0];
+          const hookIndex = this.hookIndex++;
+          const hook = this.vm.hooks[hookIndex];
+          if (hook) {
+            return [hook.state, hook.setState];
+          }
+
+          const getState = () => this.vm.hooks[hookIndex]?.state;
+
+          const setState = (newState) => {
+            if (newState instanceof Function) {
+              let currentState = getState();
+              newState = newState(currentState);
+            }
+            this.vm.setReactHook(hookIndex, { state: newState, setState });
+            return newState;
+          };
+
+          return [setState(initialState), setState];
+        } else if (callee === "useEffect") {
+          if (this.prevStack) {
+            throw new Error(
+              "Method: useEffect. The hook can an only be called from the top of the stack"
+            );
+          }
+          if (args.length < 1) {
+            throw new Error(
+              "Method: useEffect. Required arguments: 'setup'. Optional: 'dependencies'"
+            );
+          }
+          const setup = args[0];
+          if (!(setup instanceof Function)) {
+            throw new Error(
+              "Method: useEffect. The first argument 'setup' must be a function"
+            );
+          }
+          const hookIndex = this.hookIndex++;
+          const dependencies = args[1];
+          const hook = this.vm.hooks[hookIndex];
+          if (hook) {
+            const oldDependencies = hook.dependencies;
+            if (
+              oldDependencies !== undefined &&
+              deepEqual(oldDependencies, dependencies)
+            ) {
+              return undefined;
+            }
+          }
+          const cleanup = hook?.cleanup;
+          if (isFunction(cleanup)) {
+            cleanup();
+          }
+          this.vm.setReactHook(hookIndex, {
+            cleanup: setup(),
+            dependencies,
+          });
+
+          return undefined;
         } else if (callee === "setTimeout") {
           const [callback, timeout] = args;
           const timer = setTimeout(() => {
@@ -1777,7 +1849,22 @@ export default class VM {
     }
 
     this.setReactState = setReactState
-      ? (s) => setReactState(isObject(s) ? Object.assign({}, s) : s)
+      ? (s) =>
+          setReactState({
+            hooks: this.hooks,
+            state: isObject(s) ? Object.assign({}, s) : s,
+          })
+      : () => {
+          throw new Error("State is unavailable for modules");
+        };
+    this.setReactHook = setReactState
+      ? (i, v) => {
+          this.hooks[i] = v;
+          setReactState({
+            hooks: this.hooks,
+            state: this.state.state,
+          });
+        }
       : () => {
           throw new Error("State is unavailable for modules");
         };
@@ -2046,7 +2133,7 @@ export default class VM {
     );
   }
 
-  execCode({ props, context, state, forwardedProps }) {
+  execCode({ props, context, reactState, forwardedProps }) {
     if (this.compileError) {
       throw this.compileError;
     }
@@ -2054,6 +2141,8 @@ export default class VM {
       return "Too deep";
     }
     this.gIndex = 0;
+    const { hooks, state } = reactState;
+    this.hooks = hooks;
     this.state = {
       props: isObject(props) ? Object.assign({}, props) : props,
       context,