+++ CHANGELOG.md
@@ -2,6 +2,34 @@
 
 ## Pending
 
+- Introduce `useState` and `useEffect`. They should work similarly to the React hooks. Example:
+```jsx
+const [a, setA] = useState("Y");
+
+const [b, setB] = useState("B");
+const [sum, setSum] = useState(0);
+
+useEffect(() => {
+  setSum(a.length + b.length);
+  return () => {
+    console.log("cleanup");
+  };
+}, [a, b]);
+
+return (
+  <div>
+    A = {a}
+    <br />B = {b}
+    <br />
+    Length sum = {sum}
+    <div>
+      <button onClick={() => setA(a + "O")}>A</button>
+      <button onClick={() => setB(b + "O")}>B</button>
+    </div>
+  </div>
+);
+```
+
 - Add `cacheOptions` optional argument to the following methods:
   - `Social.get(keys, blockId|finality, options, cacheOptions)`
   - `Social.getr(keys, blockId|finality, options, cacheOptions)`